lettuce/pyproject.toml
[project]
name = "lettuce"
version = "0.1.0"
description = "An LLM assistant for mapping OMOP concepts"
authors = [{name = "James Mitchell White", email = "mbxjim@exmail.notingham.ac.uk"}]
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "haystack-ai>=2.7.0",
    "huggingface-hub>=0.24.6",
    "numpy==1.26.4",
    "python-dotenv>=1.0.1",
    "llama-cpp-haystack>=0.4.1",
    "llama-cpp-python>=0.2.89",
    "fastapi>=0.112.2",
    "uvicorn>=0.30.6",
    "sse-starlette>=2.1.3",
    "psycopg2-binary>=2.9.9",
    "rapidfuzz>=3.9.6",
    "SQLAlchemy>=2.0.32",
    "sseclient-py>=1.8.0",
    "fastembed-haystack>=1.2.0",
    "qdrant-haystack>=7.0.0",
    "torch==2.2",
    "sentence-transformers>=3.3.1"
]

[project.optional-dependencies]
test = ["pytest>=8.3.3"]
streamlit = ["streamlit>=1.37.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
lettuce-cli = "cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]